registrazione partita

+++ src/main/java/it/uniroma3/siw/tornei/repository/ArbitroRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import it.uniroma3.siw.tornei.model.Arbitro;
 
-public interface ArbitroRepository extends CrudRepository<Arbitro, Integer>{
+public interface ArbitroRepository extends CrudRepository<Arbitro, Long>{
 
 }
